linting and Updates for qc.data accesses

+++ src/qiskit_qec/analysis/combinations.h
@@ -23,7 +23,7 @@ class Combinations
   public:
     Combinations(int n_, int k_);
     bool next_combination(std::vector<int> &state);
-    long int size(void);
+    int size(void);
 };
 
 #endif
